+++ src/test/java/tetro/BlockTest.java
@@ -1,8 +1,8 @@
 package tetro;
 
 import org.junit.jupiter.api.Test;
-import tetro.cell.Cell;
-import tetro.cell.Cells;
+import tetro.grid.Cell;
+import tetro.grid.Cells;
 
 import static org.junit.jupiter.api.Assertions.*;
