+++ src/test/java/tetro/GridStringTest.java
@@ -2,8 +2,8 @@ package tetro;
 
 import org.junit.jupiter.api.Test;
 
-import tetro.cell.Cell;
-import tetro.cell.Cells;
+import tetro.grid.Cell;
+import tetro.grid.Cells;
 
 import static org.junit.jupiter.api.Assertions.*;
