+++ src/test/java/tetro/grid/CellTest.java
@@ -1,4 +1,4 @@
-package tetro.cell;
+package tetro.grid;
 
 import org.junit.jupiter.api.Test;
